Implementei o giro da Matriz Circular

+++ Projetos/Furos/Classes/MatrizCircular.cs
@@ -9,7 +9,7 @@ namespace Furos.Classes
     public class MatrizCircular
     {
         public double Raio { get; set; }
-        public int NumeroFuros { get; set; }
+        public double NumeroFuros { get; set; }
         public double GrauRotacao { get; set; }
 
         public MatrizCircular() { }
@@ -24,7 +24,7 @@ namespace Furos.Classes
         public MatrizCircular(string r, string nF, string gR)
         {
             Raio = double.Parse(r);
-            NumeroFuros = int.Parse(nF);
+            NumeroFuros = double.Parse(nF);
             GrauRotacao = double.Parse(gR);
         }
     }
